pyproject.toml
[project]
name = "quietsignal-backend"
version = "0.1.0"
description = "Backend for QuietSignal mood diary"
readme = "README.md"
authors = [
    { name = "Alvaro Silva", email = "102283382+BabyScylla69@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "passlib[bcrypt]",
    "PyJWT",
    "joblib",
    "scikit-learn",
    "sqlalchemy",
    "pymysql>=1.1.2",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.12.0",
]

# Tell uv/setuptools where the package lives
[tool.uv]
package = true

[tool.uv.sources]
quietsignal-backend = { path = "." }

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"
